requirements.txt
fastapi
uvicorn[standard]
openai
httpx
pydantic
pydantic-settings
beautifulsoup4
lxml
readability-lxml
python-dotenv